requirements.txt
pandas==2.2.3
pyarrow==17.0.0
pyreadstat==1.2.7
scikit-learn==1.5.2
xgboost==2.1.1
streamlit==1.51.0
apache-airflow==2.10.2
joblib==1.4.2
requests==2.32.3
